employee_management_system/requirements.txt
absl-py==0.9.0
amqp==2.6.0
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.1
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
bleach==3.1.0
blis==0.4.1
camelot-py==0.7.3
catalogue==1.0.0
celery==4.4.6
certifi==2020.12.5
chardet==4.0.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
convertdate==2.2.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
dataclasses==0.7
decorator==4.4.1
deepforest==0.2.10
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.7
django-celery==3.3.1
django-celery-results==1.2.1
django-filter==2.4.0
djangorestframework==3.12.0
djangorestframework-simplejwt==4.4.0
dlib==19.6.0
Elixir==0.7.1
en-core-web-lg==2.2.5
en-core-web-md==2.2.5
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Fiona==1.8.13
future==0.18.2
gast==0.3.3
geopandas==0.6.3
glob3==0.0.1
google-images-download==2.8.0
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
holidays==0.10.1
idna==2.10
image==1.5.28
imageio==2.6.1
importlib-metadata==3.7.2
importlib-resources==5.1.2
imutils==0.5.3
ipdb==0.12.3
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keplergl==0.1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-resnet==0.1.0
kiwisolver==1.1.0
kombu==4.6.11
lunardate==0.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
munch==2.5.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
opencv-wrapper==0.2.3
openpyxl==3.0.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
patsy==0.5.1
pbr==5.4.5
pdfminer.six==20200402
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
plotly==4.5.0
preshed==3.0.2
progress==1.5
progressbar2==3.50.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.7.0
psycopg2==2.8.4
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycocotools==2.0.0
pycryptodome==3.9.7
Pygments==2.5.2
PyJWT==1.7.1
PyMeeus==0.3.6
pyparsing==2.4.6
PyPDF2==1.26.0
pyproj==2.4.2.post1
pyrsistent==0.15.7
pystan==2.19.1.1
pytesseract==0.2.5
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
redis==3.4.1
requests==2.22.0
retrying==1.3.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
setuptools-git==1.2
Shapely==1.7.0
six==1.14.0
sklearn==0.0
slidingwindow==0.0.13
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.2.3
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
srsly==1.0.1
statsmodels==0.11.0
Tempita==0.5.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
tesserocr==2.5.0
testpath==0.4.4
textblob==0.15.3
tflearn==0.3.2
thinc==7.3.1
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
traittypes==0.2.1
Transform==0.0.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
virtualenv==20.4.3
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==0.7.1
xlwt==0.7.2
xmltodict==0.12.0
zipp==2.1.0
